# services/api-gateway/cmd/api-gateway/openapi.yaml
openapi: 3.0.3
info:
  title: golang-essential-3 API
  version: 1.0.0
  description: Public REST edge for the Signal Ledger portfolio.
servers:
  - url: /
tags: [{ name: health }, { name: auth }, { name: tasks }, { name: activity }, { name: analytics }]
paths:
  /health/live:
    get:
      tags: [health]
      responses: { '200': { description: Process is alive } }
  /health/ready:
    get:
      tags: [health]
      responses:
        '200': { description: Gateway is ready }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /healthz:
    get:
      tags: [health]
      responses: { '307': { description: Backward-compatible redirect to live health } }
  /api/v1/auth/login:
    post:
      tags: [auth]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }
      responses:
        '200': { description: Access token and refresh session, content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/RefreshRequest' } } } }
      responses:
        '200': { description: Rotated token pair, content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/logout:
    post:
      tags: [auth]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/RefreshRequest' } } } }
      responses:
        '204': { description: Refresh session revoked and cookie cleared }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/tasks:
    get:
      tags: [tasks]
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }]
      responses:
        '200': { description: Paginated tasks, content: { application/json: { schema: { $ref: '#/components/schemas/TaskPage' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [tasks]
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateTaskRequest' } } } }
      responses:
        '201': { description: Task created, content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/tasks/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
    get:
      tags: [tasks]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Task, content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [tasks]
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateTaskRequest' } } } }
      responses:
        '200': { description: Task updated, content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [tasks]
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Task deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/activities:
    get:
      tags: [activity]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Admin activity timeline, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Activity' } } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/analytics/summary:
    get:
      tags: [analytics]
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/From' }, { $ref: '#/components/parameters/To' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { description: Event summary, content: { application/json: { schema: { $ref: '#/components/schemas/AnalyticsSummary' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/analytics/timeseries:
    get:
      tags: [analytics]
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/From' }, { $ref: '#/components/parameters/To' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { description: Daily event counts, content: { application/json: { schema: { $ref: '#/components/schemas/TimeseriesResponse' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/analytics/statuses:
    get:
      tags: [analytics]
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/From' }, { $ref: '#/components/parameters/To' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { description: Task status counts, content: { application/json: { schema: { $ref: '#/components/schemas/StatusesResponse' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes: { bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT } }
  parameters:
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    From: { name: from, in: query, schema: { type: string, format: date-time } }
    To: { name: to, in: query, schema: { type: string, format: date-time } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
  responses:
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: Role or ownership policy denied, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    ServiceUnavailable: { description: Dependency unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    Error: { type: object, required: [error], properties: { error: { type: string } } }
    LoginRequest: { type: object, required: [email, password], properties: { email: { type: string, format: email }, password: { type: string, format: password } } }
    RefreshRequest: { type: object, properties: { refresh_token: { type: string } } }
    TokenResponse: { type: object, required: [access_token, token_type, expires_in, user_id, role], properties: { access_token: { type: string }, refresh_token: { type: string }, token_type: { type: string, example: Bearer }, expires_in: { type: integer }, user_id: { type: string, format: uuid }, role: { type: string, enum: [admin, member] } } }
    CreateTaskRequest: { type: object, required: [title], properties: { title: { type: string, maxLength: 200 }, description: { type: string } } }
    UpdateTaskRequest: { type: object, required: [title, status], properties: { title: { type: string, maxLength: 200 }, description: { type: string }, status: { type: string, enum: [todo, doing, done] } } }
    Task: { type: object, properties: { id: { type: string, format: uuid }, owner_id: { type: string, format: uuid }, title: { type: string }, description: { type: string }, status: { type: string }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } } }
    TaskPage: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Task' } }, page: { type: integer }, page_size: { type: integer }, total: { type: integer } } }
    Activity: { type: object, properties: { id: { type: string, format: uuid }, event_id: { type: string, format: uuid }, event_type: { type: string }, task_id: { type: string, format: uuid }, actor_id: { type: string, format: uuid }, occurred_at: { type: string, format: date-time } } }
    AnalyticsSummary: { type: object, properties: { total_events: { type: integer }, created: { type: integer }, updated: { type: integer }, deleted: { type: integer }, generated_at: { type: string, format: date-time }, data_through: { type: string, format: date-time } } }
    TimeseriesResponse: { type: object, properties: { points: { type: array, items: { type: object, properties: { day: { type: string }, created: { type: integer }, updated: { type: integer }, deleted: { type: integer } } } }, generated_at: { type: string, format: date-time }, data_through: { type: string, format: date-time } } }
    StatusesResponse: { type: object, properties: { statuses: { type: array, items: { type: object, properties: { status: { type: string }, count: { type: integer } } } }, generated_at: { type: string, format: date-time }, data_through: { type: string, format: date-time } } }